+++ src/main/scala/HackerRank/Training/Search/GridLandMetro.scala
@@ -6,14 +6,14 @@ import java.util.InputMismatchException
 import scala.collection.generic.CanBuildFrom
 import scala.collection.immutable.{NumericRange, Vector}
 import scala.collection.mutable
-import scala.language.{higherKinds, implicitConversions}
+import scala.language.higherKinds
 import scala.reflect.ClassTag
 
 /**
   * Copyright (c) 2017 A. Roberto Fischer
   *
   * @author A. Roberto Fischer <a.robertofischer@gmail.com> on 8/7/2017
-  */
+*/
 private[this] object GridLandMetro {
 
   import Reader._
@@ -177,7 +177,7 @@ private[this] object GridLandMetro {
       val sb = new java.lang.StringBuilder
       while (!isSpaceChar(b)) {
         sb.appendCodePoint(b)
-        b = readByte()
+        b = readByte().toInt
       }
       sb.toString
     }
@@ -187,12 +187,12 @@ private[this] object GridLandMetro {
       var b = 0
       var minus = false
       while ( {
-        b = readByte()
+        b = readByte().toInt
         b != -1 && !((b >= '0' && b <= '9') || b == '-')
       }) {}
       if (b == '-') {
         minus = true
-        b = readByte()
+        b = readByte().toInt
       }
       while (true) {
         if (b >= '0' && b <= '9') {
@@ -200,7 +200,7 @@ private[this] object GridLandMetro {
         } else {
           if (minus) return -num else return num
         }
-        b = readByte()
+        b = readByte().toInt
       }
       throw new IOException("Read Int")
     }
@@ -210,12 +210,12 @@ private[this] object GridLandMetro {
       var b = 0
       var minus = false
       while ( {
-        b = readByte()
+        b = readByte().toInt
         b != -1 && !((b >= '0' && b <= '9') || b == '-')
       }) {}
       if (b == '-') {
         minus = true
-        b = readByte()
+        b = readByte().toInt
       }
       while (true) {
         if (b >= '0' && b <= '9') {
@@ -223,7 +223,7 @@ private[this] object GridLandMetro {
         } else {
           if (minus) return -num else return num
         }
-        b = readByte()
+        b = readByte().toInt
       }
       throw new IOException("Read Long")
     }
@@ -232,7 +232,7 @@ private[this] object GridLandMetro {
     private[this] var lenBuffer = 0
     private[this] var ptrBuffer = 0
 
-    private[this] def readByte()(implicit in: java.io.InputStream): Int = {
+    private[this] def readByte()(implicit in: java.io.InputStream): Byte = {
       if (lenBuffer == -1) throw new InputMismatchException
       if (ptrBuffer >= lenBuffer) {
         ptrBuffer = 0
@@ -255,7 +255,7 @@ private[this] object GridLandMetro {
     private[this] def skip = {
       var b = 0
       while ( {
-        b = readByte()
+        b = readByte().toInt
         b != -1 && isSpaceChar(b)
       }) {}
       b
